Start working on a script to use GraphQL queries to retrieve open PRs.
Pulls down the PRs from the first 100 repos but doesn't sort them. Need
to get a list of _all_ open PRs, but only 100 can be queried at a
time. Also need to filter response and sort PRs.

GitHub has a Ruby GraphQL client library which is not easy to get
working against the GitHub API.

The GitHub API is pretty finicky about what the GraphQL query needs to
look like. Newline or even tabs cause it to barf and reference the v3
API.

Thanks, GitHub.